feat: implement asynchronous CV upload and evaluation process

+++ src/main/java/fpt/org/inblue/service/impl/UserServiceImpl.java
@@ -125,6 +125,9 @@ public class UserServiceImpl implements UserService {
             )
     @Transactional
     public CandidateProfile upCv(int userId, Long applicationId, MultipartFile cvFile) throws IOException {
+        if(cvFile.isEmpty()) {
+            return null;
+        }
         User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User Not Found 123"));
         if (user.getCv_public_id() != null) {
             cloudinaryService.deletePdf(user.getCv_public_id());
